# others/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FIAP Cloud Games API",
    "description": "API para gerenciamento de jogos, biblioteca de usuários e promoções",
    "contact": {
      "name": "FIAP Cloud Games Team",
      "email": "support@fiapcloudgames.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Promotion/active": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Obtém todas as promoções ativas",
        "responses": {
          "200": {
            "description": "Retorna a lista de promoções ativas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Promotion"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Promotion/{id}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Obtém uma promoção específica pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da promoção",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a promoção encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Promotion"
                }
              }
            }
          },
          "404": {
            "description": "Promoção não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Promotion"
        ],
        "summary": "Atualiza uma promoção existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da promoção a ser atualizada",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Dados atualizados da promoção",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promoção atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Promotion"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou ID não confere",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Promotion"
        ],
        "summary": "Remove uma promoção",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da promoção a ser removida",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Promoção removida com sucesso"
          },
          "404": {
            "description": "Promoção não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/api/Promotion/game/{gameId}/active": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Obtém todas as promoções ativas de um jogo específico",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID do jogo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a lista de promoções ativas do jogo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Promotion"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Promotion/game/{gameId}/discounted-price": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Calcula o preço com desconto para um jogo específico",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID do jogo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna o preço com desconto calculado",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Promotion": {
      "post": {
        "tags": [
          "Promotion"
        ],
        "summary": "Cria uma nova promoção",
        "requestBody": {
          "description": "Dados da nova promoção",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Promoção criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Promotion"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou erro na validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Promotion": {
        "required": [
          "endDate",
          "startDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "discountPercentage": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "discountAmount": {
            "maximum": 1000,
            "minimum": 0,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Informe o token JWT no formato: Bearer {seu token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}